Add passport user y admin

+++ src/Routes/index.ts
@@ -6,6 +6,8 @@ import CtrlMail from '../Controllers/ControllerMail'
 import CtrlCategory from '../Controllers/ControllerCategory'
 import CtrlCart from '../Controllers/ControllerCart'
 
+import passport from 'passport'
+
 const api = Router()
 
 
@@ -27,20 +29,20 @@ api.get('/send', CtrlMail.send)
  
 api.get('/productos', CtrlProduct.getProducts)
 api.get('/productos/:id',CtrlProduct.getProductsCod)
-api.post('/productos/agregar',  CtrlProduct.addProduct)
+api.post('/productos/agregar',passport.authenticate('admin',{ session:false }),  CtrlProduct.addProduct)
 
 api.get('/categoria/:categoria', CtrlProduct.getProductsCtg)
 
 
 api.get('/productos/:id',CtrlProduct.getProductsCod)
-api.put('/productos/:id',  CtrlProduct.updateStock)
-api.delete('/productos/:id',  CtrlProduct.deleteProduct)
+api.put('/productos/:id',passport.authenticate('admin',{ session:false }),  CtrlProduct.updateStock)
+api.delete('/productos/:id',passport.authenticate('admin',{ session:false }),  CtrlProduct.deleteProduct)
 
 api.get('/categorias', CtrlCategory.getCategory)
-api.post('/categorias', CtrlCategory.addCategory)
+api.post('/categorias/agregar',passport.authenticate('admin',{ session:false }), CtrlCategory.addCategory)
 
 
-api.get('/carro', CtrlCart.getCart)
+api.get('/carro',passport.authenticate('user',{ session:false }), CtrlCart.getCart)
 
 
 export default api;
